Uploading entire framework
This includes all the basic functions necessary to produce a product. It has functional menu switching, only the ball upgrades are functional. The rest of the menus are empty, and the settings is a WIP that I never completed. But regardless it should have most of the logic and basics to build off of.