Pushing the Framework for the Incremental Game
This is the framework for a product, all the basic logic exists. All you need to do to complete this is to finish the other menus, as the ball upgrades already exist and are in working order. The settings menu is lackluster, and the scaling of the gameobjects does not work for tablets, only works for phones. The assets, scripts, and sounds are all readily available inside the framework